@@ -13,30 +13,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.2.4",
-    "@angular/cdk": "^17.2.2",
-    "@angular/common": "^17.2.4",
-    "@angular/compiler": "^17.2.4",
-    "@angular/core": "^17.2.4",
-    "@angular/forms": "^17.2.4",
-    "@angular/material": "^17.2.2",
-    "@angular/platform-browser": "^17.2.4",
-    "@angular/platform-browser-dynamic": "^17.2.4",
-    "@angular/router": "^17.2.4",
+    "@angular/animations": "^17.3.0",
+    "@angular/cdk": "^17.3.0",
+    "@angular/common": "^17.3.0",
+    "@angular/compiler": "^17.3.0",
+    "@angular/core": "^17.3.0",
+    "@angular/forms": "^17.3.0",
+    "@angular/material": "^17.3.0",
+    "@angular/platform-browser": "^17.3.0",
+    "@angular/platform-browser-dynamic": "^17.3.0",
+    "@angular/router": "^17.3.0",
     "bootstrap": "^5.3.3",
+    "ngxtension": "^2.1.0",
     "rxjs": "^7.8.0",
     "tslib": "^2.5.0",
     "zone.js": "^0.14.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.2.3",
+    "@angular-devkit/build-angular": "^17.3.0",
     "@angular-eslint/builder": "^17.2.1",
     "@angular-eslint/eslint-plugin": "^17.2.1",
     "@angular-eslint/eslint-plugin-template": "^17.2.1",
     "@angular-eslint/schematics": "17.2.1",
     "@angular-eslint/template-parser": "^17.2.1",
-    "@angular/cli": "^17.2.3",
-    "@angular/compiler-cli": "^17.2.4",
+    "@angular/cli": "^17.3.0",
+    "@angular/compiler-cli": "^17.3.0",
     "@commitlint/cli": "^19.0.3",
     "@commitlint/config-conventional": "^19.0.3",
     "@types/jasmine": "^5.1.4",
