@@ -1,6 +1,6 @@
-import { Component, EventEmitter, Input, Output } from '@angular/core';
+import { Component, input, output } from '@angular/core';
 import { ReactiveFormsModule, UntypedFormGroup } from '@angular/forms';
-import { NgIf } from '@angular/common';
+
 import { MatInput } from '@angular/material/input';
 import { MatError, MatFormField } from '@angular/material/form-field';
 
@@ -9,10 +9,10 @@ import { MatError, MatFormField } from '@angular/material/form-field';
   templateUrl: './cart-shipping-form.component.html',
   styleUrls: ['./cart-shipping-form.component.scss'],
   standalone: true,
-  imports: [ReactiveFormsModule, MatFormField, MatInput, NgIf, MatError],
+  imports: [ReactiveFormsModule, MatFormField, MatInput, MatError],
 })
 export class CartShippingFormComponent {
-  @Input() shippingInfo!: UntypedFormGroup;
+  shippingInfo = input.required<UntypedFormGroup>();
 
-  @Output() nextStep = new EventEmitter<void>();
+  nextStep = output();
 }
