@@ -1,15 +1,13 @@
-import { Component, OnInit } from '@angular/core';
 import {
-  UntypedFormBuilder,
-  UntypedFormGroup,
-  Validators,
-} from '@angular/forms';
+  ChangeDetectionStrategy,
+  Component,
+  computed,
+  inject,
+} from '@angular/core';
+import { UntypedFormBuilder, Validators } from '@angular/forms';
 import { STEPPER_GLOBAL_OPTIONS } from '@angular/cdk/stepper';
 import { CheckoutService } from './checkout.service';
-import { ProductCheckout } from '../products/product.interface';
-import { Observable } from 'rxjs';
 import { CartService } from './cart.service';
-import { map, shareReplay } from 'rxjs/operators';
 import { CartShippingFormComponent } from './cart-shipping-form/cart-shipping-form.component';
 import { MatButton } from '@angular/material/button';
 import { MatProgressSpinner } from '@angular/material/progress-spinner';
@@ -21,7 +19,8 @@ import {
   MatStepperPrevious,
 } from '@angular/material/stepper';
 import { MatCard, MatCardContent, MatCardTitle } from '@angular/material/card';
-import { AsyncPipe, NgIf } from '@angular/common';
+import { AsyncPipe } from '@angular/common';
+import { toSignal } from '@angular/core/rxjs-interop';
 
 @Component({
   selector: 'app-cart',
@@ -35,7 +34,6 @@ import { AsyncPipe, NgIf } from '@angular/common';
   ],
   standalone: true,
   imports: [
-    NgIf,
     MatCard,
     MatCardTitle,
     MatCardContent,
@@ -49,20 +47,33 @@ import { AsyncPipe, NgIf } from '@angular/common';
     MatStepperPrevious,
     AsyncPipe,
   ],
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class CartComponent implements OnInit {
-  products$!: Observable<ProductCheckout[]>;
-  totalPrice$!: Observable<number>;
-  totalInCart$!: Observable<number>;
-  cartEmpty$!: Observable<boolean>;
+export class CartComponent {
+  private readonly fb = inject(UntypedFormBuilder);
+  private readonly checkoutService = inject(CheckoutService);
+  private readonly cartService = inject(CartService);
 
-  shippingInfo!: UntypedFormGroup;
+  products = toSignal(this.checkoutService.getProductsForCheckout(), {
+    initialValue: [],
+  });
 
-  constructor(
-    private readonly fb: UntypedFormBuilder,
-    private readonly checkoutService: CheckoutService,
-    private readonly cartService: CartService,
-  ) {}
+  totalPrice = computed(() => {
+    const products = this.products();
+    const total = products.reduce((acc, val) => acc + val.totalPrice, 0);
+    return +total.toFixed(2);
+  });
+
+  cartNotEmpty = computed(() => {
+    return this.cartService.totalInCart() > 0;
+  });
+
+  shippingInfo = this.fb.group({
+    lastName: ['', Validators.required],
+    firstName: ['', Validators.required],
+    address: ['', Validators.required],
+    comment: '',
+  });
 
   get fullName(): string {
     const { firstName, lastName } = this.shippingInfo.value;
@@ -75,36 +86,6 @@ export class CartComponent implements OnInit {
 
   get comment(): string {
     return this.shippingInfo.value.comment;
-  }
-
-  ngOnInit(): void {
-    this.shippingInfo = this.fb.group({
-      lastName: ['', Validators.required],
-      firstName: ['', Validators.required],
-      address: ['', Validators.required],
-      comment: '',
-    });
-
-    this.products$ = this.checkoutService.getProductsForCheckout().pipe(
-      shareReplay({
-        refCount: true,
-        bufferSize: 1,
-      }),
-    );
-
-    this.totalPrice$ = this.products$.pipe(
-      map((products) => {
-        const total = products.reduce((acc, val) => acc + val.totalPrice, 0);
-        return +total.toFixed(2);
-      }),
-      shareReplay({
-        refCount: true,
-        bufferSize: 1,
-      }),
-    );
-
-    this.totalInCart$ = this.cartService.totalInCart$;
-    this.cartEmpty$ = this.totalInCart$.pipe(map((count) => count > 0));
   }
 
   add(id: string): void {
