@@ -1,32 +1,28 @@
-import { Injectable } from '@angular/core';
+import { inject, Injectable } from '@angular/core';
 import { CartService } from './cart.service';
 import { ProductsService } from '../products/products.service';
 import { Observable } from 'rxjs';
 import { ProductCheckout } from '../products/product.interface';
-import { map, switchMap } from 'rxjs/operators';
+import { map } from 'rxjs/operators';
 
 @Injectable({
   providedIn: 'root',
 })
 export class CheckoutService {
-  constructor(
-    private readonly cartService: CartService,
-    private readonly productsService: ProductsService
-  ) {}
+  private readonly cartService = inject(CartService);
+  private readonly productsService = inject(ProductsService);
 
   getProductsForCheckout(): Observable<ProductCheckout[]> {
-    return this.cartService.cart$.pipe(
-      switchMap((cart) =>
-        this.productsService.getProductsForCheckout(Object.keys(cart)).pipe(
-          map((products) =>
-            products.map((product) => ({
-              ...product,
-              orderedCount: cart[product.id],
-              totalPrice: +(cart[product.id] * product.price).toFixed(2),
-            }))
-          )
-        )
-      )
+    const cart = this.cartService.cart();
+
+    return this.productsService.getProductsForCheckout(Object.keys(cart)).pipe(
+      map((products) =>
+        products.map((product) => ({
+          ...product,
+          orderedCount: cart[product.id],
+          totalPrice: +(cart[product.id] * product.price).toFixed(2),
+        })),
+      ),
     );
   }
 }
