@@ -1,5 +1,5 @@
-import { Injectable, Injector } from '@angular/core';
-import { ApiEndpoint, Config } from '../../environments/config.interface';
+import { inject, Injectable } from '@angular/core';
+import { ApiEndpoint } from '../../environments/config.interface';
 import { CONFIG_TOKEN } from './injection-tokens/config.token';
 import { Location } from '@angular/common';
 import { HttpClient } from '@angular/common/http';
@@ -7,13 +7,8 @@ import { HttpClient } from '@angular/common/http';
 /** Base class for services working with APIs */
 @Injectable()
 export abstract class ApiService {
-  protected readonly config: Config;
-  protected readonly http: HttpClient;
-
-  protected constructor(protected readonly injector: Injector) {
-    this.config = injector.get(CONFIG_TOKEN);
-    this.http = injector.get(HttpClient);
-  }
+  protected readonly config = inject(CONFIG_TOKEN);
+  protected readonly http = inject(HttpClient);
 
   endpointEnabled(api: ApiEndpoint): boolean {
     return this.config.apiEndpointsEnabled[api];
