@@ -1,11 +1,11 @@
-import { Injectable } from '@angular/core';
+import { inject, Injectable } from '@angular/core';
 import { MatSnackBar } from '@angular/material/snack-bar';
 
 @Injectable({
   providedIn: 'root',
 })
 export class NotificationService {
-  constructor(private readonly snackBar: MatSnackBar) {}
+  #snackBar = inject(MatSnackBar);
 
   /**
    * Show message with an error
@@ -14,7 +14,7 @@ export class NotificationService {
    * @param duration Duration to close after. 0 to close manually only
    */
   showError(text: string, duration = 3000) {
-    this.snackBar.open(text, 'Dismiss', {
+    this.#snackBar.open(text, 'Dismiss', {
       duration,
       panelClass: 'shop-snackbar-error',
     });
