@@ -1,13 +1,21 @@
-import { Component, ElementRef, Input, OnInit, ViewChild } from '@angular/core';
+import {
+  ChangeDetectionStrategy,
+  Component,
+  computed,
+  effect,
+  ElementRef,
+  inject,
+  Injector,
+  input,
+  viewChild,
+} from '@angular/core';
 import { Product } from '../product.interface';
 import { CartService } from '../../cart/cart.service';
-import { Observable } from 'rxjs';
-import { map, shareReplay, tap } from 'rxjs/operators';
 import { CartCountControlsComponent } from '../../core/cart-count-controls/cart-count-controls.component';
 import { MatIcon } from '@angular/material/icon';
 import { MatTooltip } from '@angular/material/tooltip';
 import { MatIconButton } from '@angular/material/button';
-import { AsyncPipe, CurrencyPipe, DecimalPipe, NgIf } from '@angular/common';
+import { AsyncPipe, CurrencyPipe, DecimalPipe } from '@angular/common';
 import {
   MatCard,
   MatCardActions,
@@ -29,7 +37,6 @@ import {
     MatCardTitle,
     MatCardContent,
     MatCardActions,
-    NgIf,
     MatIconButton,
     MatTooltip,
     MatIcon,
@@ -38,66 +45,58 @@ import {
     DecimalPipe,
     CurrencyPipe,
   ],
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class ProductItemComponent implements OnInit {
-  @Input() product!: Product;
-  @Input() index!: number;
+export class ProductItemComponent {
+  product = input.required<Product>();
+  index = input.required<number>();
 
-  @ViewChild('cartBtn', { static: false, read: ElementRef }) cartBtn:
-    | ElementRef<HTMLButtonElement>
-    | undefined;
-  @ViewChild('controls', { static: false }) countControls:
-    | CartCountControlsComponent
-    | undefined;
+  cartBtn = viewChild('cartBtn', { read: ElementRef<HTMLButtonElement> });
+  countControls = viewChild('controls', { read: CartCountControlsComponent });
 
-  countInCart$!: Observable<number>;
+  #injector = inject(Injector);
+  #cartService = inject(CartService);
 
-  constructor(private readonly cartService: CartService) {}
+  countInCart = computed(() => {
+    const cart = this.#cartService.cart();
+
+    if (!(this.id in cart)) {
+      return 0;
+    }
+
+    return cart[this.id];
+  });
 
   get id(): string {
-    return this.product.id;
+    return this.product().id;
   }
 
-  ngOnInit(): void {
-    this.countInCart$ = this.cartService.cart$.pipe(
-      map((cart) => {
-        if (!(this.id in cart)) {
-          return 0;
-        }
-
-        return cart[this.id];
-      }),
-      this.updateFocusIfNeeded(),
-      shareReplay({
-        bufferSize: 1,
-        refCount: true,
-      }),
-    );
+  constructor() {
+    this.updateFocusIfNeeded();
   }
 
   add(): void {
-    this.cartService.addItem(this.id);
+    this.#cartService.addItem(this.id);
   }
 
   remove(): void {
-    this.cartService.removeItem(this.id);
+    this.#cartService.removeItem(this.id);
   }
 
   /** Move focus to a corresponding control when controls switch */
   private updateFocusIfNeeded() {
     let prev: number;
 
-    return (observable: Observable<number>): Observable<number> =>
-      observable.pipe(
-        tap((curr) => {
-          if (prev === 0 && curr === 1) {
-            setTimeout(() => this.countControls?.focusAddBtn());
-          } else if (prev === 1 && curr === 0) {
-            setTimeout(() => this.cartBtn?.nativeElement.focus());
-          }
+    effect(() => {
+      const curr = this.countInCart();
 
-          prev = curr;
-        }),
-      );
+      if (prev === 0 && curr === 1) {
+        setTimeout(() => this.countControls()?.focusAddBtn());
+      } else if (prev === 1 && curr === 0) {
+        setTimeout(() => this.cartBtn()?.nativeElement.focus());
+      }
+
+      prev = curr;
+    });
   }
 }
