@@ -1,20 +1,18 @@
-import { Component } from '@angular/core';
+import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
 import { ProductsService } from './products.service';
-import { Observable } from 'rxjs';
-import { Product } from './product.interface';
 import { ProductItemComponent } from './product-item/product-item.component';
-import { AsyncPipe, NgFor } from '@angular/common';
+import { toSignal } from '@angular/core/rxjs-interop';
 
 @Component({
   selector: 'app-products',
   templateUrl: './products.component.html',
   styleUrls: ['./products.component.scss'],
   standalone: true,
-  imports: [NgFor, ProductItemComponent, AsyncPipe],
+  imports: [ProductItemComponent],
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class ProductsComponent {
-  readonly products$: Observable<Product[]> =
-    this.productsService.getProducts();
-
-  constructor(private readonly productsService: ProductsService) {}
+  products = toSignal(inject(ProductsService).getProducts(), {
+    initialValue: [],
+  });
 }
