@@ -1,18 +1,14 @@
-import { Injectable, Injector } from '@angular/core';
+import { Injectable } from '@angular/core';
 import { EMPTY, Observable } from 'rxjs';
 import { ApiService } from '../../core/api.service';
 import { switchMap } from 'rxjs/operators';
 
 @Injectable()
 export class ManageProductsService extends ApiService {
-  constructor(injector: Injector) {
-    super(injector);
-  }
-
   uploadProductsCSV(file: File): Observable<unknown> {
     if (!this.endpointEnabled('import')) {
       console.warn(
-        'Endpoint "import" is disabled. To enable change your environment.ts config'
+        'Endpoint "import" is disabled. To enable change your environment.ts config',
       );
       return EMPTY;
     }
@@ -24,8 +20,8 @@ export class ManageProductsService extends ApiService {
             // eslint-disable-next-line @typescript-eslint/naming-convention
             'Content-Type': 'text/csv',
           },
-        })
-      )
+        }),
+      ),
     );
   }
 
